remove code duplication;  URL maps service as constant
