+++ spinta/types/geometry/constants.py
@@ -1,3 +1,4 @@
 SRID_NOTSET = -1
 WGS84 = 4326
 LKS94 = 3346
+OPENSTREETMAP_URL = "https://www.openstreetmap.org"
